slide_generator_api/requirements.txt
fastapi
uvicorn
python-dotenv
python-pptx
google-generativeai
pydantic